pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "houston-energy-mapper"
version = "0.1.0"
description = "AI-powered venture-scale startup mapping across the Houston energy ecosystem"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    # ── LLM ──────────────────────────────────────────────────────────────────
    "anthropic>=0.40.0",

    # ── HTTP and scraping ────────────────────────────────────────────────────
    "requests>=2.32",
    "httpx>=0.27",           # async HTTP for concurrent harvesting batches
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    "playwright>=1.45",      # headless Chromium for JS-rendered pages
    "feedparser>=6.0",       # RSS (InnovationMap, EnergyCapitalHTX RSS feeds)

    # ── PDF extraction ───────────────────────────────────────────────────────
    "pdfplumber>=0.11",      # DOE award docs, corporate VC PDFs

    # ── Data handling ────────────────────────────────────────────────────────
    "pandas>=2.2",
    "openpyxl>=3.1",         # .xlsx export (deliverable spreadsheet)

    # ── Deduplication ────────────────────────────────────────────────────────
    "rapidfuzz>=3.9",        # fuzzy name matching — faster than thefuzz, no C dep issues

    # ── Utilities ────────────────────────────────────────────────────────────
    "python-slugify>=8.0",   # canonical company ID generation
    "python-dotenv>=1.0",    # .env loading
    "tenacity>=9.0",         # retry logic with exponential backoff and jitter
    "pydantic>=2.7",         # structured LLM output validation
    "jinja2>=3.1",           # prompt template rendering (safer than f-strings for conditionals)
    "pyyaml>=6.0",           # config/corporate_vc_sources.yaml and other config files

    # ── CLI and terminal output ───────────────────────────────────────────────
    "typer>=0.12",
    "rich>=13.0",            # progress bars, tables, styled run logs
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.5",
    "mypy>=1.10",
    "types-requests",
    "types-beautifulsoup4",
]

[project.scripts]
# Allows `hem` as a CLI command after `pip install -e .`
hem = "cli:app"

# ── Tool configuration ────────────────────────────────────────────────────────

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]  # line length handled by formatter

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
